refactor stable fp.tensors module using Lift
