+++ fp/tensors/__init__.py
@@ -1,15 +1,16 @@
-from .tensor import Tensor, Numpy, Jax, Torch, TorchBackend
+from .tensor_base import TensorBase
 from .shape import Torus
 from .shape import Torus as Shape
 from .tens import Tens, Linear, Otimes
-from .backend import backend, has_jax, has_torch, np, jnp, jax, torch 
+from .tensor import Numpy, Jax, Torch, Tensor
+from .backend import (has_jax, has_torch, 
+                      np, jnp, jax, torch)
 
 __all__ = [
    "Tensor",
    "Numpy",
    "Jax",
    "Torch",
-   "TorchBackend",
    "Torus",
    "Shape",
    "Tens",
