+++ fp/tensors/backend/__init__.py
@@ -1,3 +1,3 @@
 from ._pkgs import has_jax, has_torch, np, jnp, jax, torch
-from .interfaces import backend
-backend.init()
+from .wrap_alg import Backend, Interface
+from .interfaces import TorchAPI, JaxAPI, NumpyAPI
